@@ -9,8 +9,8 @@ this_dir = os.path.abspath(os.path.dirname(inspect.getfile(inspect.currentframe(
 
 SPEC = None  # This was used to support Excel 2011: '/Applications/Microsoft Office 2011/Microsoft Excel'
 
-# TODO: uncomment to run tests with json engine
-# xw.engines.active = xw.engines['json']
+# TODO: uncomment to run tests with remote engine
+# xw.engines["remote"].activate()
 
 
 class TestBase(unittest.TestCase):
